bot/requirements.txt
aiogram==3.7.0
httpx==0.27.0
python-dotenv==1.0.1
redis==5.0.4
apscheduler==3.10.4
